# webapitc2005b/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Web API Express",
        "description": "My WebApi using Node",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3001",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/getJugador/{id}": {
            "get": {
                "tags": [
                    "Jugador"
                ],
                "summary": "Encuentra un jugador por id",
                "description": "Regresa el jugador con el id solicitado",
                "operationId": "getJugador",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id del usuario a consultaar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Jugador"
                        }
                    }
                }
            }
        },
        "/api/addPlayer": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Agrega un nuevo jugador",
                "description": "Agrega un jugador a la base de datos",
                "operationId": "addPlayer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo jugador",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updatePlayer/{id}": {
            "put": {
                "tags": [
                    "Jugador"
                ],
                "summary": "Actualiza un  jugador",
                "description": "Actualiza un jugador a la base de datos",
                "operationId": "updatePlayer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id del jugador a actualizar",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevos jugador",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Jugador"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Jugador"
                        }
                    },
                    "400": {
                        "description": "Jugador Inválido"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
        "/api/deletePlayer/{player_id}": {
            "delete": {
                "tags": [
                    "Player"
                ],
                "summary": "Borra un jugador",
                "description": "Borra un jugador de la base de datos",
                "operationId": "deletePlayer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "player_id",
                        "description": "Id del jugador a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    },
                    "400": {
                        "description": "Jugador Inválido"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
...
        "/api/getQuestion/{question_id}": {
            "get": {
                "tags": [
                    "Question&Answer"
                ],
                "summary": "Encuentra una pregunta por id",
                "description": "Regresa la pregunta con el id solicitado",
                "operationId": "getQuestion",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "question_id",
                        "description": "Id de la pregunta a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Question"
                        }
                    }
                }
            }
        },
        "/api/addQuestion": {
            "post": {
                "tags": [
                    "Question&Answer"
                ],
                "summary": "Agrega una nueva pregunta",
                "description": "Agrega una pregunta a la base de datos",
                "operationId": "addQuestion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nueva pregunta",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Question"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateQuestion/{question_id}": {
            "put": {
                "tags": [
                    "Question&Answer"
                ],
                "summary": "Actualiza una pregunta",
                "description": "Actualiza una pregunta en la base de datos",
                "operationId": "updateQuestion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "question_id",
                        "description": "Id de la pregunta a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nueva pregunta",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Question"
                        }
                    },
                    "400": {
                        "description": "Pregunta Inválido"
                    },
                    "404": {
                        "description": "Pregunta no encontrado"
                    }
                }
            }
        },
        "/api/deleteQuestion/{question_id}": {
            "delete": {
                "tags": [
                    "Question&Answer"
                ],
                "summary": "Borra una pregunta",
                "description": "Borra una pregunta de la base de datos",
                "operationId": "deleteQuestion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "question_id",
                        "description": "Id de la pregunta a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Question"
                        }
                    },
                    "400": {
                        "description": "Pregunta Inválido"
                    },
                    "404": {
                        "description": "Pregunta no encontrado"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Jugador": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "puntaje": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
        "Question": {
            "type": "object",
            "properties": {
                "question_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "level_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "group_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "question_des": {
                    "type": "string"
                },
                "question_image": {
                    "type": "string"
                }
            }
        },
...
    }
}